Update b.multi_container_pods.md
When following the answer, i ran into the issue that the "box" pods already exists. We created it with the yaml file, therefore i needed to change the name of the wget-executing container to make it work. Should'nt be a problem for most of the folks preparing for the exam, but to make it "copy-pastable" i changed the name from "box" to "box-test" as we are testing our "box" pod.